requirements.txt
certifi==2019.11.28
chardet==3.0.4
ecdsa==0.15
enum-compat==0.0.3
future==0.18.2
idna==2.9
intuit-oauth==1.2.2
oauthlib==3.1.0
pip==19.0.3
pyasn1==0.4.8
pycparser==2.18
python-dateutil==2.8.1
python-jose==3.1.0
python-quickbooks==0.8.1
rauth==0.7.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
setuptools==40.8.0
simplejson==3.17.0
six==1.14.0
urllib3==1.25.8